pyproject.toml
# ABOUTME: Project config for AI Goal Coach. Managed by uv (install, run, venv).
# ABOUTME: Dependencies will be added in Phase 1; use `uv add <pkg>` or edit here.

[project]
name = "ai-goal-coach"
version = "0.1.0"
description = "Refines vague aspirations into structured SMART goals using Google ADK and Gemini."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "google-adk>=1.25.1",
    "google-genai>=1.64.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.34",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: live tests that call Gemini (require GEMINI_API_KEY); exclude with -m 'not integration'",
]

